extract attribute package
