+++ types/string.go
@@ -3,7 +3,7 @@ package types
 import (
 	"context"
 
-	tf "github.com/hashicorp/terraform-plugin-framework"
+	"github.com/hashicorp/terraform-plugin-framework/attribute"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 )
@@ -32,14 +32,14 @@ func (s StringType) Validate(_ context.Context, _ tftypes.Value) []*tfprotov6.Di
 // Description returns a practitioner-friendly explanation of the type
 // and the constraints of the data it accepts and returns. It will be
 // combined with the Description associated with the Attribute.
-func (s StringType) Description(_ context.Context, _ tf.StringKind) string {
+func (s StringType) Description(_ context.Context, _ attribute.StringKind) string {
 	return ""
 }
 
 // ValueFromTerraform returns an AttributeValue given a tftypes.Value.
 // This is meant to convert the tftypes.Value into a more convenient Go
 // type for the provider to consume the data with.
-func (s StringType) ValueFromTerraform(_ context.Context, in tftypes.Value) (tf.AttributeValue, error) {
+func (s StringType) ValueFromTerraform(_ context.Context, in tftypes.Value) (attribute.AttributeValue, error) {
 	var val String
 	if !in.IsKnown() {
 		val.Unknown = true
@@ -73,7 +73,7 @@ func (s String) ToTerraformValue(_ context.Context) (interface{}, error) {
 
 // Equal must return true if the AttributeValue is considered
 // semantically equal to the AttributeValue passed as an argument.
-func (s String) Equal(other tf.AttributeValue) bool {
+func (s String) Equal(other attribute.AttributeValue) bool {
 	o, ok := other.(String)
 	if !ok {
 		return false
